Run tf-sdk-migrator v2upgrade and first wave of issues
Migrated from github.com/hashicorp/terraform-plugin-sdk to github.com/hashicorp/terraform-plugin-sdk/v2 following along with the upgrade guide to fix compile errors
https://www.terraform.io/docs/extend/guides/v2-upgrade-guide.html

Ensure board is stored in the state with exactly the same names as the schema. SDK has become more sensitive to mismatches and will crash.
For example, `queries` is not allowed, it must be `query` and `Caption` must be `caption`.

When reading a board, set FilterCombination to "AND" if nil. The API will return nil instead of "AND", this will cause Terraform to constanly create new plans.
Was discovered by the acceptance tests, they also seem to have become more sensitive.