Force changes to recipients to recreate the trigger
This is a workaround for #123 -- instead of attempting to update
recipients in-place, which can lead to issues where the order of
recipients as returned by API responses doesn't match their ordering in
the Terraform configuration causes spurious or invalid diffs.

A better fix would be to make the "recipient" field schema.TypeSet, but
this causes issues with computed and optional values.