datasource/query_result - take query JSON directly
By taking a Query ID it was not possible to use the results of the data source without first generating an output (or similar) making use of the results.

By switching to taking Query JSON directly, Terraform is able to use the results later in the same plan as no resources need to be created first.

Note: this is a breaking change