honeycombio_column: delete column on destroy
Since the provider's inception this resource treated deletes as a 'noop' -- simply removing the column from state.

In 0.12.0 we removed the silent import of an existing column on create which has forced this gap into the light.

Given that column deletes are fully supported in the API, it felt proper to enable them here as it is assumed that the operator understands that there is removal.