correct time_range in honeycombio_trigger example
When you use the example code in `honeycombio_trigger`'s documentation,
it will fail at time of `terraform apply`.

The error returned from the Honeycomb API says that the query duration
can't be more than 4x greater than the trigger's frequency.

The default time_range on `honeycombio_query_specification` is 7200
which is 12x larger than the `frequency` on `honeycombio_trigger`.

An alternative fix would be to change the `frequency` setting, instead.
I'd be willing to go that route, and produce another PR! It felt like
maybe it's worth mentioning the `time_range` value explicitly here, but
I'm not convinced on that.

Another possible change would be to change the default `time_range`, but
that would be a breaking change. Not only because of user expectations
but also because the provider's current trigger API requires passing
query ids from immutable query objects. If the `time_range` was updated,
it would cause the `honeycombio_query` resources to attempt to update
themselves which would then fail because queries are immutable in the
Honeycomb API.