Fix minor typo in d/query_specification documentation (#369)
<!--
Thank you for contributing to the project! 💜
Please see our [OSS process
document](https://github.com/honeycombio/home/blob/main/honeycomb-oss-lifecycle-and-practices.md#)
to get an idea of how we operate.
-->

## Short description of the changes

This PR fixes a very tiny typo in the documentation for the query
specification data source that I noticed while working with the trigger
resource.

## How to verify that this has the expected result

Look at the documentation and verify that it is correct and free of
typos.