maint: prune dead code and tidy up
With the refactors going on there was some dead code, some code
that could be easult replaced by generics, and some test code
that was doing things differently than the more recent stuff.

There's likely more to do in this vein, but this was a bit of a
timeboxed effort.