refactor(client) - testsuite and constructor improvements (#416)
This moves the embedded client's test suite into the package
`client_test` and enforces a fully external test of the client. A few
functional changes needed to happen to make this possible, but it felt
like a cleaner way to test the client given it's intended purpose.

A breaking change is included in this refactor: `client.NewClient` now
takes no arguments and creates a client with the default configuration
values, while `client.NewClientWithConfig` has been added to allow
passing a custom configuration.

This felt a bit more standard (and less awkward), and will only affect
other projects pulling in the client package.