fix(r/trigger,r/burn_alert): gracefully handle recipients with notification details on import (#665)
We've had multiple reports (and finally a repro. case!) where importing
a trigger (or presumably a burn alert) which has one or more recipients
with `notification_details` causes an error on `apply` looking something
like this:

```
honeycombio_trigger.test: Modifying... [id=2TQRBJ7xp3s]
╷
│ Error: Provider produced inconsistent result after apply
│
│ When applying changes to honeycombio_trigger.test, provider "provider[\"registry.terraform.io/honeycombio/honeycombio\"]" produced an unexpected new value: .recipient: planned set element
│ cty.ObjectVal(map[string]cty.Value{"id":cty.StringVal("tqGosJVZzkN"), "notification_details":cty.ListValEmpty(cty.Object(map[string]cty.Type{"pagerduty_severity":cty.String,
│ "variable":cty.Set(cty.Object(map[string]cty.Type{"name":cty.String, "value":cty.String}))})), "target":cty.StringVal("#test.93259235823"), "type":cty.StringVal("slack")}) does not correlate
│ with any element in actual.
│
│ This is a bug in the provider, which should be reported in the provider's own issue tracker.
```

The general description of the fix is that on import we will now
populate the recipients as sparsely as possible (just `id`), and allow
later mapping between the remote side (API) and the configuration to
resolve state.

### Testing

Annoyingly, a _cannot_ get the test suite to put state in the correct,
uh, state to trigger this bug but it can be done manually by:

1. writing a configuration for a trigger with a "pagerduty" or "webhook"
recipient and applying it
1. removing that trigger from state (`terraform state rm
honeycombio_trigger.mytrigger`)
1. importing that trigger (`terraform import
honeycombio_trigger.mytrigger dataset/3hjsdf38`)
1. running `terraform apply`

Following the steps above prior to this change produces the
"inconsistent result after apply" error above.


- Closes #640 

---
- To see the specific tasks where the Asana app for GitHub is being
used, see below:
  - https://app.asana.com/0/0/1210057444771670