fix(r/trigger): updating env-wide trigger causes apply error (#709)
Updating an env-wide trigger was not handling the `dataset` attribute
correctly resulting in an apply-time error.

Also noticed that Trigger's didn't get the same update as SLOs to
support importing environment-wide triggers by 'bare ID' (e.g. no need
to use `__all__`), so fixed that here and updated the test.

- Closes #702