fix(r/slo): SLO update causing resource replacement (#710)
Updating SLO resource after recreation cause the resource to be
completely replaced. This was due to `RequiresReplace` forcing a new
update everytime.

- add a new plan modifier that only requires the SLO resource to be
replaced when `datasets` is changed
- add `UseStateForUnknown` which copies the prior state value, if not
null. This will address the bug where the SLO cannot be found during
updates

<!--
Thank you for contributing to the project! 💜
Please see our [OSS process
document](https://github.com/honeycombio/home/blob/main/honeycomb-oss-lifecycle-and-practices.md#)
to get an idea of how we operate.
-->

## Which problem is this PR solving?

-
https://github.com/honeycombio/terraform-provider-honeycombio/issues/701

## Short description of the changes

Updating SLO resource after recreation cause the resource to be
completely replaced. This was due to `RequiresReplace` forcing a new
update everytime.

- add a new plan modifier that only requires the SLO resource to be
replaced when `datasets` is changed
- add `UseStateForUnknown` which copies the prior state value, if not
null. This will address the bug where the SLO cannot be found during
updates

## How to verify that this has the expected result
1. Create a SLO resource and note the ID
2. Change any field in the SLO and note the ID
3. The ID should now be different