Accommodate the "dataset_slug" field in triggers
Augment the Honeycomb API client in the "client" package to convey the
"dataset_slug" field for triggers. When querying for triggers at the
environment level, the API returns all triggers within the
environment—including those tied to a particular dataset. Clients can
read the populated "dataset_slug" field to distinguish which triggers
are tied to which dataset and those that sit at the environment level,
above and separate from any particular dataset.

Note that this change does not yet cover the Terraform provider's
publishing to or querying of data from the Honeycomb API as exposed
through its Terraform resources and data sources.