Accommodate the "dataset" field in board panels
Augment the Honeycomb API client in the "client" package to allow
clients to discern within a board with which dataset a given query
panel's query is associated.

For boards of the newer "flexible" type, their panels that present
queries have a "dataset" field that conveys the name—crucially, not
the slug—of the dataset on which the query operates, if any. This
field is empty—not the sentinel "__all__" value—for queries that
operate at the environment level and are not tied to any particular
dataset.

Note that this change does not yet cover the Terraform provider's
publishing of data to the Honeycomb API as exposed through its
"honeycombio_flexible_board" Terraform resource.