fix(client): add missing Dataset field to QueryPanel (#724)
Adds the missing `dataset` field to the flexible board `QueryPanel`
struct.

Originally noted by @seh (with thanks!) the tests in that PR didn't
pass, so to avoid the back -- and asking more time of a community member
-- I fixed it here.

In the process noticed that some boards weren't getting cleaned up in
tests, so added a 'Sweeper' to keep that tidy.

- Replaces #723