feat(features): add features support with column import override (#731)
This introduces a `features` block to the provider configuration
allowing 'feature toggles' to be provided to resources as we see fit.
This design is heavily influenced by the Azure Provider.

The first feature toggle is included, enabling the ability for the
columns resource to automatically import a column during creation on
conflict (which, was the behaviour prior to v0.12.0 of this provider).

The features block is part of the provider configuration and is designed
to be extended per resource as we see fit in the future.

To enable the "import on conflict" behavior for the columns resource you
would initialize your provider like this:

```hcl
provider "honeycombio" {
  features {
    column {
      import_on_conflict = true
    }
  }
}
```

This partially resolves #552, and I intend to follow on with an
`r/dataset` feature toggle for the same.

Kudos to @elimt for spiking out much of this -- I've listed him as a
co-author.

---------

Co-authored-by: Elim Tsiagbey <elimtsiagbey@honeycomb.io>