fix(d/query_specification): Allow wider granularity range (#762)
<!--
Thank you for contributing to the project! 💜
Please see our [OSS process
document](https://github.com/honeycombio/home/blob/main/honeycomb-oss-lifecycle-and-practices.md#)
to get an idea of how we operate.
-->

## Which problem is this PR solving?

- Closes
https://app.asana.com/1/72322038700732/project/1201358523078372/task/1211428571544421?focus=true

## Short description of the changes

This updates the granularity validation logic to match the API, where we
allow granularities that create as few as 1 bucket in the query results,
instead of 10.

## How to verify that this has the expected result

Tests pass, I think.